SmpTsk/requirements.txt
dj-database-url==0.3.0
Django==1.8.1
django-classy-tags==0.6.1
django-crispy-forms==1.4.0
django-front==0.4.8
django-inplaceedit==1.3.0
django-scribbler==0.5.3
gunicorn==19.3.0
pluggy==0.3.0
py==1.4.27
six==1.9.0
tox==2.0.1
virtualenv==12.1.1
whitenoise==1.0.6
